System Style Revision Part 4

+++ Services/Style/System/classes/Overview/class.ilSystemStyleOverviewGUI.php
@@ -48,10 +48,6 @@ class ilSystemStyleOverviewGUI
      */
     protected $ref_id;
 
-    /**
-     * @var ilTree
-     */
-    protected $tree;
 
     /**
      * Constructor
@@ -67,7 +63,6 @@ class ilSystemStyleOverviewGUI
         $this->toolbar = $DIC->toolbar();
         $this->lng = $DIC->language();
         $this->tpl = $DIC["tpl"];
-        $this->tree = $DIC["tree"];
 
         $this->ref_id = (int) $_GET["ref_id"];
     }
@@ -87,7 +82,7 @@ class ilSystemStyleOverviewGUI
             case "saveNewSubStyle":
             case "edit":
             case "copyStyle":
-            case "delete":
+            case "deleteStyle":
             case "deleteStyles":
             case "moveUserStyles":
             case "saveStyleSettings":
@@ -221,7 +216,7 @@ class ilSystemStyleOverviewGUI
             // currently existing style
             foreach($all_user_styles as $style)
             {
-                if (ilStyleDefinition::doesStyleExist($style))
+                if (ilStyleDefinition::styleExists($style))
                 {
                     $style_arr = explode(":", $style);
                     ilObjUser::_moveUsersToStyle($style_arr[0],$style_arr[1],$to[0],$to[1]);
@@ -282,120 +277,6 @@ class ilSystemStyleOverviewGUI
         $this->ctrl->redirect($this , "edit");
     }
 
-    ////
-    //// Substyles
-    ////
-
-    /**
-     * Assign styles to categories
-     *
-     * @param
-     * @return
-     */
-    function assignStylesToCats()
-    {
-        $this->ctrl->setParameter($this, "style_id", urlencode($_GET["style_id"]));
-
-        $this->checkPermission("sty_write_system");
-
-        $all_styles = ilStyleDefinition::getAllSkinStyles();
-        $sel_style = $all_styles[$_GET["style_id"]];
-
-        $options = array();
-        if (is_array($sel_style["substyle"]))
-        {
-            foreach ($sel_style["substyle"] as $subst)
-            {
-                $options[$subst["id"]] = $subst["name"];
-            }
-        }
-
-        // substyle
-        include_once("./Services/Form/classes/class.ilSelectInputGUI.php");
-        $si = new ilSelectInputGUI($this->lng->txt("sty_substyle"), "substyle");
-        $si->setOptions($options);
-        $this->toolbar->addInputItem($si, true);
-
-        $this->toolbar->addFormButton($this->lng->txt("sty_add_assignment"), "addStyleCatAssignment");
-        $this->toolbar->setFormAction($this->ctrl->getFormAction($this));
-
-        include_once("./Services/Style/System/classes/Overview/class.ilSysStyleCatAssignmentTableGUI.php");
-        $tab = new ilSysStyleCatAssignmentTableGUI($this, "assignStylesToCats");
-
-        $this->tpl->setContent($tab->getHTML());
-    }
-
-
-    /**
-     * Add style category assignment
-     *
-     * @param
-     * @return
-     */
-    function addStyleCatAssignment()
-    {
-        $this->checkPermission("sty_write_system");
-
-        $this->ctrl->setParameter($this, "style_id", urlencode($_GET["style_id"]));
-        $this->ctrl->setParameter($this, "substyle", urlencode($_REQUEST["substyle"]));
-
-        include_once 'Services/Search/classes/class.ilSearchRootSelector.php';
-        $exp = new ilSearchRootSelector(
-            $this->ctrl->getLinkTarget($this,'addStyleCatAssignment'));
-        $exp->setExpand($_GET["search_root_expand"] ? $_GET["search_root_expand"] : $this->tree->readRootId());
-        $exp->setExpandTarget($this->ctrl->getLinkTarget($this,'addStyleCatAssignment'));
-        $exp->setTargetClass(get_class($this));
-        $exp->setCmd('saveStyleCatAssignment');
-        $exp->setClickableTypes(array("cat"));
-
-        // build html-output
-        $exp->setOutput(0);
-        $this->tpl->setContent($exp->getOutput());
-    }
-
-
-    /**
-     * Save style category assignment
-     *
-     * @param
-     * @return
-     */
-    function saveStyleCatAssignment()
-    {
-        $this->checkPermission("sty_write_system");
-
-        $this->ctrl->setParameter($this, "style_id", urlencode($_GET["style_id"]));
-
-        $style_arr = explode(":", $_GET["style_id"]);
-        ilStyleDefinition::writeSystemStyleCategoryAssignment($style_arr[0], $style_arr[1],
-            $_GET["substyle"], $_GET["root_id"]);
-        ilUtil::sendSuccess($this->lng->txt("msg_obj_modified"), true);
-
-        $this->ctrl->redirect($this, "assignStylesToCats");
-    }
-
-    /**
-     * Delete system style to category assignments
-     */
-    function deleteSysStyleCatAssignments()
-    {
-        $this->checkPermission("sty_write_system");
-
-        $this->ctrl->setParameter($this, "style_id", urlencode($_GET["style_id"]));
-        $style_arr = explode(":", $_GET["style_id"]);
-        if (is_array($_POST["id"]))
-        {
-            foreach ($_POST["id"] as $id)
-            {
-                $id_arr = explode(":", $id);
-                ilStyleDefinition::deleteSystemStyleCategoryAssignment($style_arr[0], $style_arr[1],
-                    $id_arr[0], $id_arr[1]);
-            }
-            ilUtil::sendSuccess($this->lng->txt("msg_obj_modified"), true);
-        }
-
-        $this->ctrl->redirect($this, "assignStylesToCats");
-    }
     /**
      * create
      */
@@ -411,7 +292,7 @@ class ilSystemStyleOverviewGUI
         if ($form->checkInput() )
         {
             include_once("Services/Style/System/classes/class.ilStyleDefinition.php");
-            if(ilStyleDefinition::doesSkinExist($_POST["skin_id"])){
+            if(ilStyleDefinition::skinExists($_POST["skin_id"])){
                 ilUtil::sendFailure($this->lng->txt("skin_id_exists"));
             }
             else{
@@ -599,8 +480,8 @@ class ilSystemStyleOverviewGUI
     }
 
     protected function deleteStyle(){
-        $skin_id = $_GET["style_id"];
-        $style_id = $_GET["style_ide_id"];
+        $skin_id = $_GET["skin_id"];
+        $style_id = $_GET["style_id"];
 
         $delete_form_table = new ilSystemStyleDeleteGUI();
         $container = ilSystemStyleSkinContainer::generateFromId($skin_id);
@@ -628,7 +509,11 @@ class ilSystemStyleOverviewGUI
                 try{
                     $imploded_skin_style_id = explode(":", $skin_style_id);
                     $container = ilSystemStyleSkinContainer::generateFromId($imploded_skin_style_id[0],$message_stack);
-                    $container->delete();
+                    $syle = $container->getSkin()->getStyle($imploded_skin_style_id[1]);
+                    $container->deleteStyle($syle);
+                    if(!$container->getSkin()->hasStyles()){
+                        $container->delete();
+                    }
                 }catch(Exception $e){
                     $message_stack->addMessage(new ilSystemStyleMessage($e->getMessage(),ilSystemStyleMessage::TYPE_ERROR));
                 }
@@ -745,7 +630,6 @@ class ilSystemStyleOverviewGUI
 
                     $style = new ilSkinStyleXML($_POST['sub_style_id'], $_POST['sub_style_name']);
                     $style->setSubstyleOf($parent_style_id);
-
                     $container->addStyle($style);
 
                     $this->ctrl->setParameterByClass('ilSystemStyleSettingsGUI','skin_id',$parent_skin_id);
